Update account info struct

+++ feature-wallet-impl/src/main/java/jp/co/soramitsu/feature_wallet_impl/data/network/struct/AccountData.kt
@@ -16,7 +16,7 @@ object AccountData : Schema<AccountData>() {
 object AccountInfo : Schema<AccountInfo>() {
     val nonce by uint32()
 
-    val refCount by uint8()
+    val refCount by uint32()
 
     val data by schema(AccountData)
 }
